refactor: consolidate chiplets auxiliary trace construction
