+++ processor/src/chiplets/aux_trace/virtual_table.rs
@@ -1,0 +1,60 @@
+use super::{ColMatrix, Felt, FieldElement, StarkField, Word};
+use crate::trace::LookupTableRow;
+
+// SIBLING TABLE
+// ================================================================================================
+
+/// Describes updates to the sibling table. The internal u32 values are indexes of added/removed
+/// rows in a list of rows sorted chronologically (i.e., from first added row to last).
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum SiblingTableUpdate {
+    SiblingAdded(u32),
+    SiblingRemoved(u32),
+}
+
+/// Describes a single entry in the sibling table which consists of a tuple `(index, node)` where
+/// index is the index of the node at its depth. For example, assume a leaf has index n. For the
+/// leaf's parent the index will be n << 1. For the parent of the parent, the index will be
+/// n << 2 etc.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct SiblingTableRow {
+    index: Felt,
+    sibling: Word,
+}
+
+impl SiblingTableRow {
+    pub fn new(index: Felt, sibling: Word) -> Self {
+        Self { index, sibling }
+    }
+}
+
+impl LookupTableRow for SiblingTableRow {
+    /// Reduces this row to a single field element in the field specified by E. This requires
+    /// at least 6 alpha values.
+    fn to_value<E: FieldElement<BaseField = Felt>>(
+        &self,
+        _main_trace: &ColMatrix<Felt>,
+        alphas: &[E],
+    ) -> E {
+        // when the least significant bit of the index is 0, the sibling will be in the 3rd word
+        // of the hasher state, and when the least significant bit is 1, it will be in the 2nd
+        // word. we compute the value in this way to make constraint evaluation a bit easier since
+        // we need to compute the 2nd and the 3rd word values for other purposes as well.
+        let lsb = self.index.as_int() & 1;
+        if lsb == 0 {
+            alphas[0]
+                + alphas[3].mul_base(self.index)
+                + alphas[12].mul_base(self.sibling[0])
+                + alphas[13].mul_base(self.sibling[1])
+                + alphas[14].mul_base(self.sibling[2])
+                + alphas[15].mul_base(self.sibling[3])
+        } else {
+            alphas[0]
+                + alphas[3].mul_base(self.index)
+                + alphas[8].mul_base(self.sibling[0])
+                + alphas[9].mul_base(self.sibling[1])
+                + alphas[10].mul_base(self.sibling[2])
+                + alphas[11].mul_base(self.sibling[3])
+        }
+    }
+}
