+++ processor/src/chiplets/bitwise/tests.rs
@@ -1,5 +1,5 @@
 use super::{
-    super::bus::{ChipletLookup, ChipletsBusRow},
+    super::aux_trace::{ChipletLookup, ChipletsBusRow},
     Bitwise, BitwiseLookup, ChipletsBus, Felt, StarkField, TraceFragment, Vec, A_COL_IDX,
     A_COL_RANGE, BITWISE_AND, BITWISE_AND_LABEL, BITWISE_XOR, BITWISE_XOR_LABEL, B_COL_IDX,
     B_COL_RANGE, OP_CYCLE_LEN, OUTPUT_COL_IDX, PREV_OUTPUT_COL_IDX, TRACE_WIDTH,
