+++ processor/src/chiplets/hasher/tests.rs
@@ -3,7 +3,7 @@ use super::{
     HasherState, Selectors, TraceFragment, Vec, Word, LINEAR_HASH, MP_VERIFY, MR_UPDATE_NEW,
     MR_UPDATE_OLD, RETURN_HASH, RETURN_STATE, TRACE_WIDTH,
 };
-use crate::chiplets::virtual_table::{AuxTraceBuilder, SiblingTableRow, SiblingTableUpdate};
+use crate::chiplets::aux_trace::{SiblingTableRow, SiblingTableUpdate, TableTraceBuilder};
 use miden_air::trace::chiplets::hasher::{
     DIGEST_LEN, HASH_CYCLE_LEN, LINEAR_HASH_LABEL, MP_VERIFY_LABEL, MR_UPDATE_NEW_LABEL,
     MR_UPDATE_OLD_LABEL, NUM_ROUNDS, NUM_SELECTORS, RETURN_HASH_LABEL, RETURN_STATE_LABEL,
@@ -66,8 +66,8 @@ fn hasher_permute() {
     assert_eq!(trace.last().unwrap(), &[ZERO; 8]);
 
     // make sure aux hints for sibling table are empty
-    assert!(aux_hints.sibling_hints.is_empty());
-    assert!(aux_hints.sibling_rows.is_empty());
+    assert!(aux_hints.sibling_hints().is_empty());
+    assert!(aux_hints.sibling_rows().is_empty());
 
     // --- test two permutations ----------------------------------------------
 
@@ -104,8 +104,8 @@ fn hasher_permute() {
     assert_eq!(trace.last().unwrap(), &[ZERO; 16]);
 
     // make sure aux hints for sibling table are empty
-    assert!(aux_hints.sibling_hints.is_empty());
-    assert!(aux_hints.sibling_rows.is_empty());
+    assert!(aux_hints.sibling_hints().is_empty());
+    assert!(aux_hints.sibling_rows().is_empty());
 }
 
 // MERKLE TREE TESTS
@@ -180,8 +180,8 @@ fn hasher_build_merkle_root() {
     assert_eq!(&node_idx_column[9..], &[ZERO; 7]);
 
     // make sure aux hints for sibling table are empty
-    assert!(aux_hints.sibling_hints.is_empty());
-    assert!(aux_hints.sibling_rows.is_empty());
+    assert!(aux_hints.sibling_hints().is_empty());
+    assert!(aux_hints.sibling_rows().is_empty());
 
     // --- Merkle tree with 8 leaves ------------------------------------------
 
@@ -222,8 +222,8 @@ fn hasher_build_merkle_root() {
     check_merkle_path(&trace, 0, leaves[5], &path, 5, MP_VERIFY);
 
     // make sure aux hints for sibling table are empty
-    assert!(aux_hints.sibling_hints.is_empty());
-    assert!(aux_hints.sibling_rows.is_empty());
+    assert!(aux_hints.sibling_hints().is_empty());
+    assert!(aux_hints.sibling_rows().is_empty());
 
     // --- Merkle tree with 8 leaves (multiple branches) ----------------------
 
@@ -338,8 +338,8 @@ fn hasher_build_merkle_root() {
     check_merkle_path(&trace, 72, leaves[3], &path3, 3, MP_VERIFY);
 
     // make sure aux hints for sibling table are empty
-    assert!(aux_hints.sibling_hints.is_empty());
-    assert!(aux_hints.sibling_rows.is_empty());
+    assert!(aux_hints.sibling_hints().is_empty());
+    assert!(aux_hints.sibling_rows().is_empty());
 }
 
 #[test]
@@ -447,11 +447,11 @@ fn hasher_update_merkle_root() {
         (16, SiblingTableUpdate::SiblingAdded(1)),
         (24, SiblingTableUpdate::SiblingRemoved(1)),
     ];
-    assert_eq!(expected_sibling_hints, aux_hints.sibling_hints);
+    assert_eq!(expected_sibling_hints, aux_hints.sibling_hints());
 
     let expected_sibling_rows =
         vec![SiblingTableRow::new(ZERO, path0[0]), SiblingTableRow::new(ONE, path1[0])];
-    assert_eq!(expected_sibling_rows, aux_hints.sibling_rows);
+    assert_eq!(expected_sibling_rows, aux_hints.sibling_rows());
 
     // --- Merkle tree with 8 leaves ------------------------------------------
 
@@ -608,7 +608,7 @@ fn hasher_update_merkle_root() {
         (128, SiblingTableUpdate::SiblingRemoved(7)),
         (136, SiblingTableUpdate::SiblingRemoved(8)),
     ];
-    assert_eq!(expected_sibling_hints, aux_hints.sibling_hints);
+    assert_eq!(expected_sibling_hints, aux_hints.sibling_hints());
 
     let expected_sibling_rows = vec![
         // first update
@@ -624,7 +624,7 @@ fn hasher_update_merkle_root() {
         SiblingTableRow::new(Felt::new(3 >> 1), path3_2[1]),
         SiblingTableRow::new(Felt::new(3 >> 2), path3_2[2]),
     ];
-    assert_eq!(expected_sibling_rows, aux_hints.sibling_rows);
+    assert_eq!(expected_sibling_rows, aux_hints.sibling_rows());
 }
 
 // MEMOIZATION TESTS
@@ -1038,7 +1038,7 @@ fn hash_memoization_span_blocks_check(span_block: CodeBlock) {
 
 /// Builds an execution trace for the provided hasher. The trace must have the number of rows
 /// specified by num_rows.
-fn build_trace(hasher: Hasher, num_rows: usize) -> (Vec<Vec<Felt>>, AuxTraceBuilder) {
+fn build_trace(hasher: Hasher, num_rows: usize) -> (Vec<Vec<Felt>>, TableTraceBuilder) {
     let mut trace = (0..TRACE_WIDTH).map(|_| vec![Felt::new(0); num_rows]).collect::<Vec<_>>();
     let mut fragment = TraceFragment::trace_to_fragment(&mut trace);
     let aux_trace_builder = hasher.fill_trace(&mut fragment);
